+++ src/main/java/com/sape/safety_for_people/repository/SaleRepository.java
@@ -3,5 +3,5 @@ package com.sape.safety_for_people.repository;
 import com.sape.safety_for_people.model.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleRepository extends JpaRepository<Sales, Integer> {
+public interface SaleRepository extends JpaRepository<Sales, Long> {
 }
